+++ packages/eyes-webdriverio-5/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- deprecate `saveDebugData`
 
 ## 5.25.1 - 2020/10/15